+++ src/components/App/Router/index.js
@@ -16,7 +16,8 @@ import ListingList from '../../Listing';
 import ListingDisplay from '../../Listing/ListingDisplay';
 import Forgot from "../../Auth/Forgot";
 import AddBio from "../../Profile/AddBio";
-import Match from "../../Match/Match";
+import Match from "../../Match";
+import MatchDetails from "../../Match/MatchDetails";
 import BottomNav from "../Navigation/BottomNav";
 import { useLocation } from "react-router";
 
@@ -66,6 +67,7 @@ const Router = () => {
     <ProtectedRoute exact path="/listing/display" component={ListingDisplay} />
     <ProtectedRoute exact path="/profile" component={Profile} />
     <ProtectedRoute exact path="/match" component={Match} />
+    <ProtectedRoute exact path="/match/details" component={MatchDetails} />
     <Route component={ErrorRoute} />
   </Switch>
   {setDisplay && <BottomNav/>}
